backend/pyproject.toml
[project]
name = "csv-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "fastmcp>=0.1.0",
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "langchain>=0.2.0",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-experimental>=0.3.4",
    "langchain-mcp-adapters>=0.1.0",
    "langchain-openai>=0.1.0",
    "langgraph>=0.1.0",
    "matplotlib>=3.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pathlib2>=2.3.7 ; python_full_version < '3.4'",
    "plotly>=5.15.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.10.0",
    "seaborn>=0.12.0",
    "streamlit>=1.49.1",
    "tabulate>=0.9.0",
    "uvicorn>=0.24.0",
]
